Bug fixes + cast bar

+++ Senior Project 1/Assets/Scripts/HPManager.cs
@@ -59,8 +59,8 @@ public class HPManager : MonoBehaviour
 
     void Update()
     {
-        health.fillAmount = PlayerController.Player_HP / PlayerController.Player_MaxHP;
-        XP.fillAmount = PlayerController.Player_EXP / PlayerController.Player_MaxEXP;
+        health.fillAmount = (float)PlayerController.Player_HP / PlayerController.Player_MaxHP;
+        XP.fillAmount = (float)PlayerController.Player_EXP / PlayerController.Player_MaxEXP;
         healthText.text = PlayerController.Player_HP.ToString() + " / " + PlayerController.Player_MaxHP.ToString();
         XPText.text = PlayerController.Player_EXP.ToString() + " / " + PlayerController.Player_MaxEXP.ToString();
         LevelText.text = "Level: " + PlayerController.Player_Level.ToString();
